Code cleanup and enable all working assignments in CMAKE
